# docs/sad/03-design/openapi.yaml
openapi: 3.0.3
info:
  title: JPLearn Platform API
  version: 0.1.0
  description: 'Foundation v1. No grammar, flashcards, or L1 translation channels.


    Contract authority is this file (OpenAPI 3.0.3), not FastAPI-generated 3.1 JSON.

    Semantic normalized diff (ADR-003 D9): must match HTTP status, required/nullability,

    security, operationId, x-jplearn-fr, and error schema. Allowlist: 3.0↔3.1 nullable

    vs type union, generated component names, servers, info.


    Docs UI (`/docs`, `/redoc`, `/openapi.json`): off on staging and prod; local dev
    may

    enable. Contract tests read this file in git — do not depend on a public OpenAPI
    URL.

    '
servers:
- url: https://api.staging.jplearn.local
paths:
...
  /auth/register:
    post:
      operationId: register
      tags:
      - Identity
      x-jplearn-fr:
      - FR-ID-001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 10
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
  /auth/login:
    post:
      operationId: login
      tags:
      - Identity
      x-jplearn-fr:
      - FR-ID-001
      - FR-ID-002
      - NFR-SEC-003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Http401Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
  /auth/logout:
    post:
      operationId: logout
      tags:
      - Identity
      x-jplearn-fr:
      - FR-ID-003
      security:
      - bearerAuth: []
      responses:
        '204':
          description: Invalidates every access_token for this user (all devices);
            tokenVersion increment (FR-ID-003)
        '401':
          description: Missing or invalid Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Http401Error'
  /me:
    get:
      operationId: getMe
      tags:
      - Identity
      x-jplearn-fr:
      - FR-ID-002
      - FR-ID-004
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /staff/content-jobs/{id}/attempts:
    get:
      tags:
      - Staff AI Attempts
      operationId: listStaffAiAttempts
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AiAttemptPublic'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Http401Error'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
      x-jplearn-fr:
      - FR-AI-001
  /staff/content-jobs/{id}/attempts/{attempt_id}/reconcile:
    post:
      tags:
      - Staff AI Attempts
      operationId: reconcileStaffAiAttempt
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: attempt_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconcileAttemptRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiAttemptPublic'
        '400':
          description: Invalid reconciliation evidence or usage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Http401Error'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Attempt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '409':
          description: Attempt state or evidence conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
      x-jplearn-fr:
      - FR-AI-001
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    signedQuery:
      type: apiKey
      in: query
      name: sig
      description: HMAC-SHA256 signature for media streaming (with exp timestamp)
  schemas:
...
    HttpError:
      type: object
      description: Standard JSON error body (400, 403, 404, 500) per ADR-005.
      required:
      - statusCode
      - message
      - error
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
          - type: string
          - type: array
            items:
              type: string
        error:
          type: string
    Http401Error:
      type: object
      description: Distinct 401 Unauthorized JSON error body per ADR-005.
      required:
      - statusCode
      - message
      properties:
        statusCode:
          type: integer
          enum:
          - 401
        message:
          type: string
...
    AuthSession:
      type: object
      required:
      - access_token
      - user
      properties:
        access_token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required:
      - id
      - email
      - roles
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        roles:
          type: array
          items:
            type: string
            enum:
            - learner
            - teacher
            - admin
...
    AttemptUsage:
      type: object
      additionalProperties: false
      required:
      - audio_seconds
      - input_tokens
      - output_tokens
      - cost_micros
      properties:
        audio_seconds:
          type: integer
          minimum: 0
          maximum: 2147483647
        input_tokens:
          type: integer
          minimum: 0
          maximum: 2147483647
        output_tokens:
          type: integer
          minimum: 0
          maximum: 2147483647
        cost_micros:
          type: integer
          minimum: 0
          maximum: 2147483647
        currency:
          type: string
          const: USD
          default: USD
    ReconcileAttemptRequest:
      type: object
      additionalProperties: false
      required:
      - decision
      - evidence
      properties:
        decision:
          type: string
          enum:
          - billed
          - not_billed
        evidence:
          type: string
          minLength: 10
          maxLength: 2000
        provider_request_id:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 200
          - type: 'null'
        usage:
          anyOf:
          - $ref: '#/components/schemas/AttemptUsage'
          - type: 'null'
    AiAttemptPublic:
      type: object
      additionalProperties: false
      required:
      - id
      - job_id
      - attempt_number
      - provider
      - idempotency_key
      - state
      - lease_expires_at
      - created_at
      - updated_at
      properties:
        id:
          type: string
        job_id:
          type: string
        attempt_number:
          type: integer
          minimum: 1
        provider:
          type: string
        idempotency_key:
          type: string
        state:
          type: string
          enum:
          - running
          - outcome_unknown
          - settled
          - not_billed
        lease_expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        provider_request_id:
          anyOf:
          - type: string
          - type: 'null'
        usage:
          anyOf:
          - type: object
            additionalProperties: true
          - type: 'null'
        evidence:
          anyOf:
          - type: string
          - type: 'null'
        resolved_by:
          anyOf:
          - type: string
          - type: 'null'
        resolution_hash:
          anyOf:
          - type: string
          - type: 'null'
...